move FunctionInfo to a different file
Cython miscompiles the dataclass
